# chore(provider-aws): improve canonicalHostedZone handling