--- a/provider/aws/aws.go
+++ b/provider/aws/aws.go
@@ -1161,11 +1161,14 @@ func isAWSAlias(ep *endpoint.Endpoint) string {
 
 // canonicalHostedZone returns the matching canonical zone for a given hostname.
 func canonicalHostedZone(hostname string) string {
-	parts := strings.Split(hostname, ".")
-	for i := 0; i < len(parts); i++ {
-		suffix := strings.Join(parts[i:], ".")
-		if zone, exists := canonicalHostedZones[suffix]; exists {
-			return zone
+	// strings.HasSuffix is optimized for this specific task and avoids the overhead associated with compiling and executing a regular expression.
+	if strings.HasSuffix(hostname, "aws.com") || strings.HasSuffix(hostname, "aws.com.cn") || strings.HasSuffix(hostname, "tor.com") || strings.HasSuffix(hostname, "ont.com") || strings.HasSuffix(hostname, "ont.net") {
+		parts := strings.Split(hostname, ".")
+		for i := len(parts) - 2; i >= 0; i-- {
+			suffix := strings.Join(parts[i:], ".")
+			if zone, exists := canonicalHostedZones[suffix]; exists {
+				return zone
+			}
 		}
 	}