<!--
    Please read https://github.com/kubernetes-sigs/external-dns#contributing before submitting
    your pull request. Please fill in each section below to help us better prioritize your pull request. Thanks!
-->

**Description**

<!-- Please provide a summary of the change here. -->

<!-- Please link to all GitHub issue that this pull request implements(i.e. Fixes #123) -->
No issue to date.

Just a small improvements  to how canonicalHostedZone handled.



Current number of known canonical hosted zones is 130+ and is increasing, in worth case scenario current logic does 130+ iteration to validate where or not the record is a canonical hosted zone. 
```
results for canonical

before
BenchmarkTestAWSCanonicalHostedZone-16    	    7012	    171223 ns/op

now
BenchmarkTestAWSCanonicalHostedZone-16    	   16812	     68977 ns/op

results for long non-canonical hosted zone

before
enchmarkTestAWSNonCanonicalHostedZone-16    	    4869	    225384 ns/op

with new algo
BenchmarkTestAWSNonCanonicalHostedZone-16    	  277828	      4135 ns/op
```
Comparison
- Algorithm: `canonicalHostedZoneNew uses a more targeted approach by first filtering hostnames based on specific suffixes, while canonicalHostedZoneOld checks all possible suffixes in the map.
- Efficiency: This approach is optimized for hostnames with specific suffixes, as it avoids unnecessary checks for other domains. The use of strings.HasSuffix is efficient for this task vs regex pattern matching.
- Use Case: `canonicalHostedZoneN` might be more efficient if the `canonicalHostedZonesOld` map is getting large, as it reduces the number of lookups. `canonicalHostedZoneOld`  might be more efficient for a more general solution. This new function is suitable when you have a predefined set of suffixes and you want to quickly find the canonical hosted zone for hostnames matching those suffixes.

Another improvement could be: to create a cache, so that it checked hostname once and store a result, but this require additional memory proportional to the number of hostnames.

**Checklist**

- [x] Unit tests updated
- [ ] End user documentation updated